ExFmt: Implement injector system
This commit implements the first part of the new exchange-format (exfmt)
and its conversion-system. Specifically, this commit implements the
injector system responsible for injecting objects into the new format.

A future commit will be responsible for the extractor-system, used to
extract objects from the new format. For testability, this commit
includes some temporary (and ugly) extractor implementations.