Fix two bugs concerning determinism and the StreetGraph's implementation
The StreetGraph implemented a HashSet containing all nodes and edges. We wanted a HashMap from a node's or edge's ID to its reference. The problem was the fact that two edges, forward and backward, have the same ID. Therefore the map didn't work correctly. This has been fixed by using the hashcode as key. In case an edge has an identical origin and destination node, a new value 'Orientation' is added containing "FORWARD" and "BACKWARD", influencing the hashcode.

All test cases are passing.