Change main-ui debugging settings; multilane-info below:
Maps with a lot of traffic jams does not have many jams anymore. Only remaining jams are recognized at crossroads, where two lanes (full of vehicles) are laying next to each other. Maybe the reason is the fact that nodes are prefering vehicles waiting longer for passing the crossroad, but the prefered vehicle has to change its lane before it can pass the crossroad.