Fix deadlock at the end of roads
It is possible that two vehicles wait for each other at the end of a
road. To prevent this, a `critical zone` has been introduced directly in
front of a crossroad. The implementation guarantees no deadlocks because
only certain vehicles enter the critical zone.

Only lane changes lead the vehicles to wait for each other on the same
road. Hence vehicles, which are already driving on the correct lane,
will leave the critical zone without trouble. They can enter it in every
case.

Vehicles, which still have to change their lane before crossing the
crossroad, are only allowed to enter the critical zone if (1) it is
empty. Without further criterion, the deadlock would still exist just
one cell in front of the critical zone. Thus (2) a vehicle is allowed to
enter the critical zone if it is the outermost vehicle on this road
(containing >= 1 lanes).