Fix phenomenom of traffic jam inversion
This means that vehicles tend to overtake and thus, the inner lanes of
a road are getting fuller (=> smaller velocity) than outer lanes.

The used model extends the NaSch-model by letting vehicles decide to
overtake after accelerating. In the decision process, the current
velocity of the front vehicle is checked for zero. After accelerating,
the velocity is not zero. Hence overtaking in traffic jams was not
possible.

Fix: Ask for velocity in the previous NaSch-step.